Add Check protocol
Makes it possible to specify how to check a certain ID type for status.

This is so that a heartbeat() function can be specified for any ID that
a user decides to make.

This means that the :type key is now split out and this is determined by
datatype instead. Thus, %KOTL.ID.{Process,Node} have been made. When
adding ID types you'll add a struct and define how to get the status for
that ID type. See the defimpl `KOTL.Monitor.Check` part in the code for
%KOTL.ID.Process{} and %KOTL.ID.Node{} for how to do this.