Add diff request and diff_current
Diff call is for getting a whole map of diffed statuses. It should be
used primarily for backend stuff. Diff with current status will be
appropriate for when %Status{} structs are needed, as they are simply
status objects. This will be good for a web view that just wants an
updated status object.